Equipment condition scripts now include context of award. NEU, cloak moved to use condition scripts rather than TL:99 hack Base equipment now uses condition script to allow OXP writers to deny particular items at their own stations. (e.g. no fuel)
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@5514 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/PlanetEntity.h
@@ -110,8 +110,8 @@ typedef struct
 - (void) setPlanetType:(OOStellarBodyType) pt;
 
 
-- (double) radius;	// metres
-- (void) setRadius:(double) rad;
+- (GLfloat) radius;	// metres
+- (void) setRadius:(GLfloat) rad;
 - (double) rotationalVelocity;
 - (void) setRotationalVelocity:(double) v;
